# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "paopao-ce API",
    "description": "# paopao-ce API 文档\n\n欢迎使用 paopao-ce API 文档！\n\npaopao-ce 是一个开源的泡泡社区系统，提供完整的社交媒体功能。\n\n## 鉴权说明\n\n大部分API需要JWT Token进行鉴权，请在请求头中包含：\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n## 响应格式\n\n所有API响应都遵循以下格式：\n```json\n{\n  \"code\": 0,\n  \"message\": \"success\",\n  \"data\": {}\n}\n```\n\n## 错误码说明\n\n- `0`: 成功\n- `10001`: 未授权\n- `10002`: 参数错误\n- `10003`: 权限不足\n- `10004`: 资源不存在\n\n## 分页参数\n\n支持分页的API使用以下查询参数：\n- `page`: 页码 (默认: 1)\n- `page_size`: 每页数量 (默认: 20)\n",
    "version": "1.0.0",
    "contact": {
      "name": "paopao-ce",
      "url": "https://github.com/rocboss/paopao-ce"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8008",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.paopao.info",
      "description": "生产环境"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/post/star": {
      "get": {
        "tags": ["动态管理"],
        "summary": "获取动态点赞状态",
        "description": "获取指定动态的点赞状态，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**请求参数：**\n- `id`: 动态ID",
        "operationId": "tweetStarStatus",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "动态ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TweetStarStatusResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      },
      "post": {
        "tags": ["动态管理"],
        "summary": "点赞动态",
        "description": "对指定动态进行点赞操作，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 重复点赞会取消点赞",
        "operationId": "starTweet",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StarTweetReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StarTweetResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "动态不存在"
          }
        }
      }
    },
    "/post/collection": {
      "get": {
        "tags": ["动态管理"],
        "summary": "获取动态收藏状态",
        "description": "获取指定动态的收藏状态，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**请求参数：**\n- `id`: 动态ID",
        "operationId": "tweetCollectionStatus",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "动态ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TweetCollectionStatusResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      },
      "post": {
        "tags": ["动态管理"],
        "summary": "收藏动态",
        "description": "对指定动态进行收藏操作，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 重复收藏会取消收藏",
        "operationId": "collectionTweet",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionTweetReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionTweetResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "动态不存在"
          }
        }
      }
    },
...
    "/attachment": {
      "post": {
        "tags": ["动态管理"],
        "summary": "上传附件",
        "description": "上传文件附件，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 支持图片、视频等多种文件类型\n- 文件大小有限制",
        "operationId": "uploadAttachment",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadAttachmentReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadAttachmentResp"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或文件格式不支持"
          },
          "401": {
            "description": "未授权"
          }
        }
      },
      "get": {
        "tags": ["动态管理"],
        "summary": "下载附件",
        "description": "下载附件文件，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**请求参数：**\n- `id`: 附件ID\n\n**注意事项：**\n- 下载前需要先调用预检接口",
        "operationId": "downloadAttachment",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "附件ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "下载成功",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限下载"
          },
          "404": {
            "description": "文件不存在"
          }
        }
      }
    },
    "/attachment/precheck": {
      "get": {
        "tags": ["动态管理"],
        "summary": "下载附件预检",
        "description": "下载附件前的权限预检，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**请求参数：**\n- `id`: 附件ID",
        "operationId": "downloadAttachmentPrecheck",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "附件ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "预检通过",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadAttachmentPrecheckResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限下载"
          }
        }
      }
    },
    "/post": {
      "post": {
        "tags": ["动态管理"],
        "summary": "发布动态",
        "description": "发布新的动态，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 动态内容不能为空\n- 支持文本、图片、视频等多种内容",
        "operationId": "createTweet",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTweetReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发布成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTweetResp"
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/post/{id}": {
      "delete": {
        "tags": ["动态管理"],
        "summary": "删除动态",
        "description": "删除指定的动态，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 只能删除自己的动态\n- 删除后不可恢复",
        "operationId": "deleteTweet",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "动态ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限删除"
          },
          "404": {
            "description": "动态不存在"
          }
        }
      }
    },
    "/post/lock": {
      "post": {
        "tags": ["动态管理"],
        "summary": "锁定动态",
        "description": "锁定指定动态，禁止评论，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 只能锁定自己的动态",
        "operationId": "lockTweet",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockTweetReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockTweetResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限操作"
          }
        }
      }
    },
...
    "/tweet/comment/thumbsup": {
      "post": {
        "tags": ["评论管理"],
        "summary": "点赞评论",
        "description": "对指定评论进行点赞，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 重复点赞会取消点赞",
        "operationId": "thumbsUpTweetComment",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TweetCommentThumbsReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "评论不存在"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "TweetStarStatusResp": {
        "type": "object",
        "properties": {
          "is_star": {
            "type": "boolean",
            "description": "是否已点赞"
          },
          "count": {
            "type": "integer",
            "description": "点赞数量"
          }
        }
      },
...
      "StarTweetReq": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "动态ID"
          }
        }
      },
      "StarTweetResp": {
        "type": "object",
        "properties": {
          "is_star": {
            "type": "boolean",
            "description": "是否已点赞"
          }
        }
      },
...
    }
  }
}